Move creatable resource types to /models
