+++ src/functions/resources/create-one-resource.spec.ts
@@ -6,10 +6,10 @@
  * MIT license. See the LICENSE file for details.
  **************************************************************************/
 
-import { isResource } from '../../models';
+import { CreatableResource, isResource } from '../../models';
 import { integrationTest } from '../../tests';
 import { TEST_AUTH_TOKEN, TEST_HOST } from '../../tests/config';
-import { CreatableResource, makeCreateOneResource } from './create-one-resource';
+import { makeCreateOneResource } from './create-one-resource';
 import { makeGetOneResource } from './get-one-resource';
 import { makeDeleteOneResource } from './delete-one-resource';
