+++ src/functions/resources/set-one-resource-content.node.spec.ts
@@ -8,10 +8,10 @@
 
 import { createReadStream, ReadStream } from 'fs';
 import { join } from 'path';
-import { isResource, Resource } from '../../models';
+import { CreatableResource, isResource, Resource } from '../../models';
 import { integrationTest } from '../../tests';
 import { TEST_ASSETS_PATH, TEST_AUTH_TOKEN, TEST_HOST } from '../../tests/config';
-import { CreatableResource, makeCreateOneResource } from './create-one-resource';
+import { makeCreateOneResource } from './create-one-resource';
 import { makeDeleteOneResource } from './delete-one-resource';
 import { makeGetOneResourceContent } from './get-one-resource-content';
 import { makeSetOneResourceContent } from './set-one-resource-content';
