+++ src/models/resource/creatable-resource.ts
@@ -6,11 +6,14 @@
  * MIT license. See the LICENSE file for details.
  **************************************************************************/
 
-export * from './creatable-resource';
-export * from './is-blank-raw-resource';
-export * from './is-resource';
-export * from './raw-creatable-resource';
-export * from './raw-resource';
-export * from './resource';
-export * from './to-raw-creatable-resource';
-export * from './to-resource';
+import { NumericID } from '../../value-objects';
+
+export interface CreatableResource {
+	groupIDs?: Array<NumericID>;
+
+	name: string;
+	description: string;
+	labels?: Array<string>;
+
+	isGlobal?: boolean;
+}
